Extract 16334 repro
